fix plot resizing
Issue:
QwtPlot sets a quite large minimum height of the plotting area.
With three or more vertical plots it is not possible to maximize
the window e.g. on a FullHD screen.

Solution:
To override and decrease the default sizeHint() and minimumSizeHint()
of QwtPlot, a new class inheriting from QwtPlot has to be created and
used in the PlotWidget.ui file.