Make your library usable
I cannot use this plugin library in my project because of the lines 304 and 305 in the CMakeLists.txt.

I am developing an ROS C++ Plugin, [rqt_alliance](https://github.com/adrianohrl/rqt_alliance), for monitoring the running, parameter-based, [ALLIANCE](https://github.com/adrianohrl/alliance) architecture for Multi-Robot Task Allocation problem solving.

By selecting a pair of robot and task (i.e. a behaviour set of a robot), it is desired to visualize multiple plots (via its respective topics) so that its motivation parametrization may be improved. Since the rqt_multiplot Plugin has the capability of doing that, my intention is to simplify it.